Separate a pivot cache records reader
